Fix test script
we had to force the new flag -summary to false, otherwise the
outputs would be always different.

For some reason if you don't pass -log flag gg will 'randomly' not
find any result.

I discovered this because I was running this script several times
and each time it would fail with a different combination of options
and queries. This would happen even if I didn't change the code
between runs.

During the debug session I tried many solutions but nothing was
effective. Two that I think should be noted are; 1) I tried to add
a sleep statement between calls to gg but I was still facing the
same issue. 2) I tried to execute the gg command again (while [gg
output is null] try again) but this just resulted in an endless
loop.

The bahaviour of not finding anything happened both in my local
branch 'addTests' and in 'master'. Sometimes my binary would not
find any reslts and sometimes binary build from master would not
find anything.

This behaviour gets even stranger because I could not reproduce it
while typing the command in an interactive shell session. It only
happened to me while gg was being called from a script.