Change signature of setupModeGG()
This change was made so it was easier to test the
function setupModeGG.

This change had the potential to alter the programs behaviour
because flags.Arg(n) returns an empty string when n does not
exist. This is why we added a little loop to fill the args
variable when get this situation.

We also made setupModeGG return a searchMode struct, and also
slightly simplified the value matcher logic.