Add multiReader struct
When we started to add tests for the logic related to actual
file reading we noticed some code duplication and slight
inconsistencies in this source.

In an attempt to fight that we came up with the idea of a
multiReader, which is an struct that acts like an interface.
This was needed because every file format you want to read
has it's own edge cases. With this we hope to bring all this
mess into a single place, and hopefully this will be better.

In this commit we didn't make the change yet, we are only
adding a struct and some tests for it.